requirements.txt
Django==1.7.8
django-debug-toolbar==1.3.0
django-extensions==1.5.5
django-filter==0.10.0
django-mptt==0.7.3
djangorestframework==3.1.2
djangorestframework-jwt==1.5.0
Pillow==2.8.1
psycopg2==2.6
PyJWT==1.3.0
six==1.9.0
sqlparse==0.1.15
wheel==0.24.0
